% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{get_marginal_stats}
\alias{get_marginal_stats}
\title{Marginal statistics for continuous trait}
\usage{
get_marginal_stats(X, y)
}
\arguments{
\item{X}{Genotype matrix.}

\item{y}{Phenotype vector.}
}
\value{
List with \code{beta} and \code{z}.
}
\description{
Marginal statistics for continuous trait
}
