% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{get_marginal_stats_bin}
\alias{get_marginal_stats_bin}
\title{Rao score marginal statistics for binary trait}
\usage{
get_marginal_stats_bin(X, y)
}
\arguments{
\item{X}{Genotype matrix.}

\item{y}{Binary phenotype vector.}
}
\value{
List with \code{beta} and \code{z}.
}
\description{
Rao score marginal statistics for binary trait
}
