% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{make_pheno}
\alias{make_pheno}
\title{Simulate phenotype}
\usage{
make_pheno(X, beta, trait, seed)
}
\arguments{
\item{X}{Genotype matrix.}

\item{beta}{True effect-size vector.}

\item{trait}{Either \code{"continuous"} or \code{"binary"}.}

\item{seed}{Random seed.}
}
\value{
Phenotype vector.
}
\description{
Simulate phenotype
}
