% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{prune_ld}
\alias{prune_ld}
\title{Greedy LD pruning}
\usage{
prune_ld(R, thresh)
}
\arguments{
\item{R}{LD correlation matrix.}

\item{thresh}{r-squared pruning threshold.}
}
\value{
Logical vector indicating retained SNPs.
}
\description{
Greedy LD pruning
}
