% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quiet}
\alias{quiet}
\title{Suppress stdout and messages}
\usage{
quiet(expr)
}
\arguments{
\item{expr}{Expression to evaluate.}
}
\value{
The value of \code{expr}.
}
\description{
Evaluates an expression while suppressing both stdout and message output.
}
