% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasSS-package.R
\docType{package}
\name{rasSS-package}
\alias{rasSS}
\alias{rasSS-package}
\title{rasSS: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/anson-li8/rasSS}
  \item \url{https://anson-li8.github.io/rasSS}
  \item Report bugs at \url{https://github.com/anson-li8/rasSS/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
