% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{scan_ss}
\alias{scan_ss}
\title{Summary-statistic RAS scan}
\usage{
scan_ss(
  b_disc,
  z_targ,
  R,
  mask,
  skip1 = 10,
  skip2 = 3,
  min_window_size = 3,
  max_window_size = 30
)
}
\arguments{
\item{b_disc}{Discovery marginal effect-size weights.}

\item{z_targ}{Target marginal Z-scores.}

\item{R}{LD correlation matrix.}

\item{mask}{Logical SNP mask.}

\item{skip1}{Pivotal SNP step size.}

\item{skip2}{Window-size step size.}

\item{min_window_size}{Minimum half-window size.}

\item{max_window_size}{Maximum half-window size.}
}
\value{
List with \code{x} and \code{y}.
}
\description{
Summary-statistic RAS scan
}
