% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_genotypes}
\alias{sim_genotypes}
\title{Simulate genotypes from an LD matrix}
\usage{
sim_genotypes(n, R)
}
\arguments{
\item{n}{Number of individuals.}

\item{R}{LD correlation matrix.}
}
\value{
Integer genotype matrix with values 0, 1, or 2.
}
\description{
Simulate genotypes from an LD matrix
}
