% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burden.R
\name{t_burden}
\alias{t_burden}
\title{Summary-statistic weighted burden statistic}
\usage{
t_burden(w, Z, R)
}
\arguments{
\item{w}{Weight vector.}

\item{Z}{Marginal Z-score vector.}

\item{R}{LD correlation matrix.}
}
\value{
Numeric burden statistic.
}
\description{
Summary-statistic weighted burden statistic
}
\examples{
w <- c(1, 1); Z <- c(1, 1); R <- matrix(c(1, 0.5, 0.5, 1), 2, 2)
t_burden(w, Z, R)
}
